split static content into another spec
